Refactor main script into modules
