+++ wkey/faster_whisper_Mother_of_all_wkey.py
@@ -9,31 +9,22 @@
 """
 
 import os
-import io
 import time
 import threading
 import pyautogui
-import winsound
-import clipboard
 import numpy as np
 import sounddevice as sd
 import pythoncom
-from scipy.io.wavfile import write as wav_write
-
-import groq
-from groq import Groq
-
 import queue
 from pynput.keyboard import Controller as KeyboardController, Key, Listener
 from dotenv import load_dotenv
 
-from faster_whisper import WhisperModel
 from voice_commands import execute_command
 from pause_all import is_sound_playing_windows_processing
+from wkey.volume_manipulation import decrease_volume_all, restore_volume_all
+from wkey.beep_utils import beep, START_BEEP, STOP_BEEP, PASTE_BEEP
+from wkey.transcription_utils import process_audio_async, clean_transcript
 
-from ctypes import cast, POINTER
-from comtypes import CLSCTX_ALL
-from pycaw.pycaw import AudioUtilities, IAudioEndpointVolume
 
 
 from vosk import Model, KaldiRecognizer
@@ -42,7 +33,6 @@ import pvporcupine
 
 
 # Initial setup and global variables
-initial_volume = None  # Variable to store initial volume
 transcript_queue = queue.Queue()
 audio_buffer_queue = queue.Queue()
 
@@ -54,9 +44,6 @@ recording = False
 stream = None
 audio_buffer = np.array([], dtype="float32")
 sample_rate = 8000
-model = WhisperModel("small.en", device="cuda", num_workers=8)
-groq_model = "distil-whisper-large-v3-en"
-# "whisper-large-v3"
 play_pause_pressed = False
 something_is_playing = False
 
@@ -87,11 +74,6 @@ wake_stream = p.open(
 wake_stream.start_stream()
 
 
-# Define beep sounds
-START_BEEP = (880, 100)  # Frequency in Hz, Duration in ms
-STOP_BEEP = (440, 100)  # Lower frequency for stop
-PASTE_BEEP = (660, 100)  # Intermediate frequency for paste
-
 # Locks for synchronization
 recording_lock = threading.Lock()
 audio_data_lock = threading.Lock()
@@ -106,35 +88,6 @@ audio_data_lock = threading.Lock()
   ## ##   ##     ## ##          ##     ## ##     ## ##    ## 
    ###     #######  ########    ##     ## ########   ######  
 """
-
-
-def get_current_volume():
-    devices = AudioUtilities.GetSpeakers()
-    interface = devices.Activate(IAudioEndpointVolume._iid_, CLSCTX_ALL, None)
-    volume = cast(interface, POINTER(IAudioEndpointVolume))
-    return volume.GetMasterVolumeLevelScalar()
-
-
-def set_volume(volume_level):
-    devices = AudioUtilities.GetSpeakers()
-    interface = devices.Activate(IAudioEndpointVolume._iid_, CLSCTX_ALL, None)
-    volume = cast(interface, POINTER(IAudioEndpointVolume))
-    volume.SetMasterVolumeLevelScalar(volume_level, None)
-
-
-def decrease_volume_all():
-    global initial_volume
-    if initial_volume is None:
-        initial_volume = get_current_volume()
-    print(f"Decreasing volume from {initial_volume * 100}% to 10%")
-    set_volume(0.1)  # Set volume to 10%
-
-
-def restore_volume_all():
-    global initial_volume
-    if initial_volume is not None:
-        print(f"Restoring volume to {initial_volume * 100}%")
-        set_volume(initial_volume)  # Restore to initial volume
 
 
 def callback(indata, frames, time, status):
@@ -304,99 +257,12 @@ def cleanup():
 """
 
 
-api_key = os.getenv("GROQ_API_KEY")
-client = Groq(api_key=api_key)
-
-
-def transcribe_with_groq(audio_buffer):
-    try:
-        # Convert the audio buffer (NumPy array) to a byte stream in WAV format
-        byte_io = io.BytesIO()
-        wav_write(byte_io, sample_rate, audio_buffer)
-        byte_io.seek(0)  # Rewind to the beginning of the byte stream
-
-        # Send the byte stream directly to the Groq API
-        transcription = client.audio.transcriptions.create(
-            file=("audio_buffer.wav", byte_io.read()),  # Use in-memory byte stream
-            model=groq_model,
-            prompt="Specify context or spelling",
-            response_format="json",
-            language="en",
-            temperature=0.0,
-        )
-        return transcription.text
-    except Exception as e:
-        print(f"Groq API error: {e}")
-        raise
-
-
-def transcribe_with_local_model(audio_buffer):
-    segments, info = model.transcribe(
-        audio_buffer, language="en", suppress_blank=True, vad_filter=True
-    )
-    transcript = " ".join([segment["text"] for segment in segments])
-    return transcript
-
-
-def process_audio_async():
-    while True:
-        try:
-            audio_buffer_for_processing = audio_buffer_queue.get(timeout=5)
-            if audio_buffer_for_processing is None:
-                break
-            try:
-                transcript = transcribe_with_groq(audio_buffer_for_processing)
-            except groq.RateLimitError:
-                print("Groq API rate limit reached, switching to local transcription.")
-                transcript = transcribe_with_local_model(audio_buffer_for_processing)
-            transcript_queue.put(transcript)
-            print(transcript)
-        except queue.Empty:
-            continue
-        except Exception as e:
-            print(f"An error occurred during transcription: {e}")
-
-
 def start_listener():
     try:
         with Listener(on_press=on_press, on_release=on_release) as listener:
             listener.join()
     except KeyboardInterrupt:
         print("Ctrl+C pressed. Exiting...")
-
-
-def beep(sound):
-    frequency, duration = sound
-    # Create and start a new thread for playing the sound
-    # thread = threading.Thread(target=lambda: winsound.Beep(frequency, duration))
-    # thread.start()
-    winsound.Beep(frequency, duration)
-    # Optionally join the thread if you want to wait for it to complete
-    # thread.join()
-
-
-# Function to clean transcript and paste it
-def clean_transcript():
-    while True:
-        try:
-            transcript = transcript_queue.get()
-
-            original_clipboard_content = clipboard.paste()
-
-            clipboard.copy(transcript)
-
-            pyautogui.hotkey("ctrl", "v")
-            beep(PASTE_BEEP)
-            print("Transcript pasted")
-
-            time.sleep(0.1)
-            clipboard.copy(original_clipboard_content)
-
-            time.sleep(0.1)
-            clipboard.copy(transcript)
-
-        except Exception as e:
-            print(f"An error occurred in clean_transcript: {e}")
 
 
 """
@@ -416,8 +282,8 @@ def main():
 
     try:
         # threading.Thread(target=monitor_sound_processing, daemon=True).start()
-        threading.Thread(target=clean_transcript, daemon=True).start()
-        threading.Thread(target=process_audio_async, daemon=True).start()
+        threading.Thread(target=clean_transcript, args=(transcript_queue,), daemon=True).start()
+        threading.Thread(target=process_audio_async, args=(audio_buffer_queue, transcript_queue, sample_rate), daemon=True).start()
         threading.Thread(target=listen_for_wake_word, daemon=True).start()
 
         with stream:
